para interp

See merge request fun3d-developers/refine!204

+++ src/ref_import.c
@@ -1569,7 +1569,7 @@ static REF_STATUS ref_import_meshb(REF_GRID *ref_grid_ptr, REF_MPI ref_mpi,
          fread((unsigned char *)&ref_geom_cad_data_size(ref_geom), 4, 1, file),
          "cad data size");
     if (verbose)
-      printf("cad_data %zd bytes\n", ref_geom_cad_data_size(ref_geom));
+      printf("cad_data %ld bytes\n", (long)ref_geom_cad_data_size(ref_geom));
     /* safe non-NULL free, if already allocated, to prevent memory leaks */
     ref_free(ref_geom_cad_data(ref_geom));
     ref_malloc_size_t(ref_geom_cad_data(ref_geom),
